use u32 instead of size_t so the merge will be cleaner

+++ EntityTest/CSlotMap.hpp
@@ -9,7 +9,7 @@
 
 #include "CEntity.hpp"
 
-template< size_t _Size, typename T >
+template< u32 _Size, typename T >
 class CSlotMap
 {
 public:
@@ -139,7 +139,7 @@ public:
 	{
 		if( nullIndex != m_lastObjectIndex )
 		{
-			for( size_t i = 0; i <= m_lastObjectIndex; i++ )
+			for( u32 i = 0; i <= m_lastObjectIndex; i++ )
 			{
 				const auto &entity = m_entities[ i ];
 				m_idMappings[ entity.Id() ] = nullIndex;
@@ -175,7 +175,7 @@ public:
 	{
 		if( nullIndex != m_lastObjectIndex )
 		{
-			for( size_t i = 0; i <= m_lastObjectIndex; i++ )
+			for( u32 i = 0; i <= m_lastObjectIndex; i++ )
 			{
 				lambda( m_entities[ i ], &m_objects[ i ] );
 			}
@@ -188,7 +188,7 @@ public:
 		if( nullIndex != m_lastObjectIndex )
 		{
 			#pragma omp parallel
-			for( size_t i = 0; i <= m_lastObjectIndex; i++ )
+			for( u32 i = 0; i <= m_lastObjectIndex; i++ )
 			{
 				lambda( m_entities[ i ], &m_objects[ i ] );
 			}
@@ -200,7 +200,7 @@ public:
 	{
 		if( nullIndex != m_lastObjectIndex )
 		{
-			for( size_t i = 0; i <= m_lastObjectIndex; i++ )
+			for( u32 i = 0; i <= m_lastObjectIndex; i++ )
 			{
 				if( lambda( m_entities[ i ], &m_objects[ i ] ) )
 				{
@@ -212,7 +212,7 @@ public:
 		return( false );
 	}
 
-	size_t Count() const
+	u32 Count() const
 	{
 		return( m_lastObjectIndex + 1 );
 	}
@@ -221,12 +221,12 @@ public:
 	const size_t SizeInBytes;
 
 private:
-	const size_t nullIndex = std::numeric_limits< size_t >::max();
+	const u32 nullIndex = std::numeric_limits< u32 >::max();
 
-	std::vector< size_t > m_idMappings;
+	std::vector< u32 > m_idMappings;
 
 	std::vector< CEntity >	m_entities;
 	std::vector< T >		m_objects;
 
-	size_t m_lastObjectIndex = nullIndex;
+	u32 m_lastObjectIndex = nullIndex;
 };
